neo_solidity::codegen

Struct CodeGenerator

pub struct CodeGenerator {
    _config: CompilerConfig,
    loop_stack: Vec<LoopContext>,
    variable_table: Vec<VariableInfo>,
    next_var_index: usize,
}

Code generator for Yul AST to NeoVM bytecode

Fields
_config: CompilerConfigloop_stack: Vec<LoopContext>

Stack of loop contexts for break/continue resolution

variable_table: Vec<VariableInfo>

Variable table for tracking local variables

next_var_index: usize

Current variable index

Implementations

impl CodeGenerator

pub fn new() -> Self

Create a new code generator with default loop stack

pub fn with_config(config: &CompilerConfig) -> Self

Create a new code generator with config

pub fn push_loop(&mut self, ctx: LoopContext)

Push a loop context onto the stack

pub fn pop_loop(&mut self) -> Option<LoopContext>

Pop a loop context from the stack

pub fn current_loop(&self) -> Option<&LoopContext>

Get the current loop context (for break/continue)

pub fn register_variable(&mut self, name: &str) -> usize

Register a new variable and return its index

pub fn get_variable_index(&self, name: &str) -> Option<usize>

Get variable index by name

pub fn reset_variables(&mut self)

Reset variable table for new function

impl CodeGenerator

pub fn generate( &mut self, ast: &AstNode, ) -> Result<CompilationResult, CompilerError>

impl CodeGenerator

fn generate_node( &mut self, node: &AstNode, bytecode: &mut Vec<u8>, functions: &mut Vec<FunctionMeta>, events: &mut Vec<String>, estimated_gas: &mut u64, ) -> Result<(), CompilerError>

impl CodeGenerator

fn generate_assembly_representation(&self, bytecode: &[u8]) -> String

impl CodeGenerator

fn generate_source_map(&self, ast: &AstNode) -> String

fn estimate_node_bytecode_size(&self, node: &AstNode) -> usize

Estimate the bytecode size a node will produce, mirroring the emission logic in node_gen.rs so source map offsets track real bytecode positions.

fn visit_ast_for_source_map( &self, node: &AstNode, source_map: &mut String, offset: usize, )

fn count_ast_nodes(&self, node: &AstNode) -> usize