neo_solidity

Module frontend

Solidity frontend integration using solang-parser.

This module parses Solidity source code into a light-weight intermediate representation that can be consumed by later compiler stages.

Structs

ContractIR

Representation of a Solidity contract.

EnumIR

EventIR

Representation of a Solidity event.

EventParameterIR

Representation of a Solidity event parameter.

FunctionIR

Representation of a Solidity function or constructor.

NatspecDocIR

Natspec documentation extracted from source comments.

ParameterIR

Representation of a Solidity parameter.

StateVariableIR

Representation of a state variable.

StructFieldIR

StructIR

Version 🔒

Enums

Bound 🔒

ComparatorOp 🔒

ContractKind

Classification of contract kinds.

FrontendError

Errors emitted by the frontend while parsing Solidity code.

MutabilityKind

Function mutability classification based on Solidity state mutability.

VisibilityKind

Functions

branch_supports_0_8 🔒

build_comment_map 🔒

Build a map from source positions to their preceding Natspec comments.

clean_doc_comment 🔒

Remove comment delimiters and leading asterisks/slashes

convert_contract 🔒

convert_enum 🔒

convert_event 🔒

convert_function 🔒

convert_parameters 🔒

convert_state_variable 🔒

convert_struct 🔒

enforce_supported_pragma 🔒

extract_mutability 🔒

extract_visibility 🔒

find_preceding_doc 🔒

Find the doc comment that precedes a given source location

format_diagnostics 🔒

function_name 🔒

intersects_0_8 🔒

next_patch 🔒

offset_to_line_column 🔒

parse_caret 🔒

parse_hyphen_range 🔒

parse_natspec 🔒

Parse Natspec tags from a documentation comment

parse_operator_version 🔒

parse_plain_version 🔒

parse_source

Parse Solidity source into ContractIR values.

parse_tilde 🔒

pragma_supports_0_8 🔒

save_tag_content 🔒

split_comparators 🔒

storage_to_string 🔒